// services/projects/src/projects/dto/update-project.dto.d.ts
import { ProjectStatus } from './create-project.dto';
export declare class UpdateProjectDto {
    title?: string;
    status?: ProjectStatus;
    startDate?: string;
    endDate?: string;
    actualEnd?: string;
    budget?: number;
    description?: string;
    metadata?: Record<string, any>;
}
//# sourceMappingURL=update-project.dto.d.ts.map